+++ core/threshold/src/execution/tfhe_internals/lwe_bootstrap_key_generation.rs
@@ -168,7 +168,7 @@ where
 
 /// Generates a Bootstrapping Key given a Glwe key in Glwe format
 /// , a Lwe key and the params for the BK generation
-#[instrument(name="Gen BK", skip(glwe_secret_key_share, lwe_secret_key_share, mpc_encryption_rng, session, preprocessing), fields(sid = ?session.session_id(), my_role = ?session.my_role()))]
+#[instrument(name="Gen BK", skip_all, fields(sid = ?session.session_id(), my_role = ?session.my_role()))]
 pub(crate) async fn generate_bootstrap_key<
     Z: BaseRing,
     P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
@@ -207,7 +207,7 @@ where
 
 /// Generates a compressed Bootstrapping Key given a Glwe key in Glwe format
 /// , a Lwe key and the params for the BK generation
-#[instrument(name="Gen compressed BK", skip(glwe_secret_key_share, lwe_secret_key_share, mpc_encryption_rng, session, preprocessing, seed), fields(sid = ?session.session_id(), my_role = ?session.my_role()))]
+#[instrument(name="Gen compressed BK", skip_all, fields(sid = ?session.session_id(), my_role = ?session.my_role()))]
 pub(crate) async fn generate_compressed_bootstrap_key<
     Z: BaseRing,
     P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
