+++ core/threshold/src/execution/tfhe_internals/sns_compression_key_generation.rs
@@ -64,7 +64,7 @@ where
     Ok(packing_key_switching_key_shares)
 }
 
-#[instrument(name="Gen Sns Compression Key", skip(glwe_secret_key_share_sns_as_lwe, private_key, mpc_encryption_rng, session, preprocessing), fields(sid = ?session.session_id(), my_role = ?session.my_role()))]
+#[instrument(name="Gen Sns Compression Key", skip_all, fields(sid = ?session.session_id(), my_role = ?session.my_role()))]
 pub(crate) async fn generate_sns_compression_keys<
     Z: BaseRing,
     P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
@@ -102,7 +102,7 @@ where
     Ok(compression_key)
 }
 
-#[instrument(name="Gen compressed Sns Compression Key", skip(glwe_secret_key_share_sns_as_lwe, private_key, mpc_encryption_rng, session, preprocessing, seed), fields(sid = ?session.session_id(), my_role = ?session.my_role()))]
+#[instrument(name="Gen compressed Sns Compression Key", skip_all, fields(sid = ?session.session_id(), my_role = ?session.my_role()))]
 pub(crate) async fn generate_compressed_sns_compression_keys<
     Z: BaseRing,
     P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
